RAG_small_tutor/backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
